flask_backend/requirements.txt
flask
flask-cors
openai
python-dotenv
openai-function-tokens
# transformers[torch]
anthropic
langdetect
python-docx
htmldocx

# test requirements
pandas
openpyxl